requirements.txt
Flask==3.0.3
openai
google-generativeai
python-dotenv
pydantic
typing